FACES-2452 Split the Liferay Faces multi-module Maven project into independently versioned Maven projects (temporary fixes for compile failures against the Liferay Portal 7.0 API)

+++ demos/portal/jsf2-registration-portlet/src/main/java/com/liferay/faces/demos/service/RegistrantServiceUtil.java
@@ -190,8 +190,8 @@ public class RegistrantServiceUtil {
 		return administratorPermissionChecker;
 	}
 
-	private static int getMobilePhoneTypeId() throws SystemException {
-		int phoneTypeId = 0;
+	private static long getMobilePhoneTypeId() throws SystemException {
+		long phoneTypeId = 0;
 		List<ListType> phoneTypes = ListTypeServiceUtil.getListTypes(PHONE_CLASS_NAME);
 
 		for (ListType phoneType : phoneTypes) {
